Refactor run module into app package
